Extracted PBO and FBO classes from the LwjglRenderingProcess class. All other changes are consequential.
